Make background blur proportional to the input video size
The blur value was a fixed value that did not change depending on the
input video size. Due to this the background of lower resolution videos
looked more blurry than the background of higher resolution videos, and
this was specially noticeable when the same video changed its
resolution. Now the blur is proportional to the input video size, so the
blurred background always looks the same even if the resolution changes
(provided the same aspect ratio is kept between the different
resolutions).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>