Fix pending device not requested once current request finishes
The changed device id handler expects the MediaDevicesManager and the
device id, but only the device id was given.

As the method was recursively called from the handler of the previous
"getUserMedia" promise trying to access "getUserMedia()" on a string
failed silently; the error was caught by the rejected promise handler
and the output track was stopped and removed.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>