Fix devices not released when got after the dialog was closed
Requesting a stream is an asynchronous operation, and it may take a
while to be fulfilled (for example, if the user needs to grant the
browser permission to use the device). When the Talk settings and the
device checker dialogs are closed the streams being used are stopped,
but this will have no effect if the promises that returns the streams
are fulfilled after that. In that case the streams need to be explicitly
stopped once finally got.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>