Fix null track temporary set when disabling the virtual background
When an output track was stopped it was automatically removed. When the
virtual background is disabled the input track is bypassed to the
output, but first the effect is stopped. This caused a null track to be
temporary set as the output track before setting the input track which,
in turn, caused different issues in other parts of the code (like
causing the video to hang in the device checker due to the stream being
removed, disabling the video during calls when disabling the background
blur due to the null stream being handled as "no video", or reconnecting
when disabling the video during calls if background blur was enabled due
to a loop in the event handling which caused the events to be handled in
the wrong order).

Even if the robustness of that other code could be improved conceptually
it makes more sense that disabling the virtual background immediately
bypasses the input track, without setting a null track first. Therefore
now the automatic removal of the output track can be disabled if needed,
and it is used when stopping the virtual background effect.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>