Revert "Use something more stable so we don't toggle the chatIdentifier when the participant list is reloaded"
This reverts commit 7d61dad0ddf2c1908cbee7b1cd61328ca0291d17.

"chatIdentifier" is used to start getting the messages, or cancel an
existing request. "isParticipant" needs to be taken into account in the
property, as it is one of the preconditions checked in
"handleStartGettingMessagesPreconditions".

Moreover, "isParticipant" should not be switching between "false" and
"true" when the participant list is updated; this is a regression
introduced in a different commit that will be fixed afterwards.