Revert "Await adding the participant before updating the user status"
This reverts commit afd9e069249a3bed9e6dfcde73f4aa47d4779339 (but
keeping the extra details in the logs).

"purgeParticipantsStore" and "addParticipant" are synchronous
operations, so there is no need to wait for them. Doing so causes the
removal of the participants and adding each participant again to happen
in a different "tick", so Vue computed properties are updated to
intermediate values (for example, the current participant switches from
being in the conversation to not being in the conversation and back
again).

This seemed to fix an issue when updating the user status, but only
because the issue needs a specific sequence of steps to trigger it in a
deterministic way, so probably it simply did not happen when it was
tested.