Fix unsubscribing listener for status updates
The listener for status updates was not unsubscribed when the
"CurrentParticipants" component was destroyed, so the component instance
was leaked and kept forever. Moreover, the listener was called again
whenever the status of a participant was updated, but as the computed
properties are no longer updated once a component is destroyed it called
"updateUser" with whichever token it had when it was destroyed rather
than the current one, so this caused "Participant not found" errors to
be printed in the console.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>