Fix enabling background blur when input track is disabled
If the background blur is enabled when the input track is disabled
the effect should not be started, as otherwise the output track will be
enabled and the effect will be just blurring a black stream.

Note that if the input track is disabled it will be already bypassed to
the output, so it should not be set again either.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>