Stop sending offers again for own peer if current session ID changed
When an own peer is needed offers are periodically sent until the
connection is established. However, if the current session ID changed
(which would happen after a forced reconnection) the previous session
ID is no longer relevant, so the offers with the old ID should not be
sent anymore (the process to send offers with the new ID will be started
when needed).

Moreover, the previous process must be stopped as well to prevent
creating new Peer objects with the old ID, which would cause them to be
added to the CallParticipantCollection (as the ID does not match with
the current session ID, so it is seen as a peer for another participant)
and thus appear in the call view.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>