Clear own peer from LocalCallParticipanModel when ended
"iceConnectionStateChanged" is not emitted when on transitions to the
"closed" state, as events are not emitted on user initiated actions.
Therefore, the own peer needs to be explicitly removed from the
LocalCallParticipanModel when ended.

In most cases this was already done, but sometimes the own peer could be
ended like a remote peer in the "disconnectedSessionIds" loop in
"usersChanged" (for example, after a forced reconnection, depending on
the timing of the signaling messages). When that happened the peer was
not removed from LocalCallParticipanModel, so it was still treated as it
was connected even if the connection was already closed (so, for
example, the analyzer still tried to get the stats and Firefox printed
"RTCPeerConnection is gone (did you enter Offline mode?)" on each try).

Due to this now "peerEnded" is emitted by the Peer object and the own
peer is removed from LocalCallParticipanModel by handling that event.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>