Show loading spinner on local video if needed even if own peer is null
The own peer and, therefore, the peer in LocalCallParticipanModel could
be null in some cases even if an own peer is needed to send media to the
HPB. This should be just a transient situation, although it could happen
too by bugs in the code.

Due to this the call view should show the local participant as "not
connected" not only when the own peer is null, but also if the own peer
is null and an own peer would be nevertheless needed.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>