Start publishing if needed on any "users changed" event
The own peer was created if needed when the list of participants in the
call actually changed. In most cases this worked fine, but sometimes
(for example, with forced reconnections, depending on the timing of the
signaling messages) it could happen that the own peer was stopped and,
once the participant list was updated, the participants were still the
same as before and thus the own peer was not started again.

Now the own peer is started if needed on any "users changed" event; this
should make no difference in the cases in which the own peer was not
needed or it was already active, but it should fix the corner cases in
which it was needed but the participant list changed without the
participants actually changing.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>