Play received screenshare audio in its own element
Similarly to the regular received audio, the audio of the received
screenshare needs to be played in its own audio element. Otherwise, if
the audio is only played by the Screen component if the component is
removed (for example, by following another user and hidding the bottom
bar) the audio would not be heard.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>