Fix "switch-to-screen-id" event
Since the LocalMediaControls were moved to the TopBar that component is
no longer a descendant of CallView, so the "switch-to-screen-id" event
is not propagated up to it. Therefore it is replaced by an event bus
event, which can be listened from the CallView.

Moreover, the "switch-to-screen-id" events emitted by other components
is not correctly propagated either up to the CallView. For consistency
and simplicity the event is moved to an event bus event in those cases
too.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>