Make possible to block the video of a receiver peer
Remote video can be blocked and allowed again by adjusting the direction
of the video transceiver of RTCPeerConnections, which triggers a
"negotiationneeded" event. When a negotiation is needed a new exchange
of offers and answers needs to be made between the peers.

Renegotiations need to be supported by both peers, so currently this is
available only with the HPB, and only on versions that support the
"update-sdp" feature.

When the HPB is being used "negotiationneeded" event causes a new offer
to be requested, which will be an update if the HPB supports it. Updates
rather than new connections are needed to prevent an audio interruption
while the connection changes.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>