Fix screenshare not stopped in receiver participants with HPB
When the HPB is used, the "unshareScreen" message was sent only if the
own screen peer existed when "localScreenStopped" was handled. However,
the own screen peer is cleared when ended, which happens before the
handler for "localScreenStopped" was executed. Due to this the signaling
message was not sent and thus the other participants did not stop the
received screenshare.

Sending the message only if the own screen peer existed was just a
legacy check which is not really needed, so now the message is always
sent when "localScreenStopped" is handled.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>