Do not show the video as disconnected while updating the connection
If the connection transitions to "disconnected" (or was disconnected
before starting the renegotiation) the UI will wrongly show the message
about reconnecting due to connection problems.

I have not found an easy way to solve this, as it would require to keep
track of whether a renegotiation started due to handling the
"negotiationneeded" event or something else (and, in the future, once
renegotiations are implemented also without HPB, it should be necessary
to know too if the offer was sent by the remote peer for other reasons).
Due to this, for now a slightly misleading message will be shown if the
connection is updated while "disconnected".

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>