Update connections rather than create new ones on "negotiationneeded"
When an offer is requested to the HPB the old connection is stopped and
a new one is created. If the HPB has support for updating the
subscribers this is now used instead when negotiation is needed, as this
prevents the existing connection (and thus the media) to be interrupted
during the renegotiation.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>