Reparent file picker and viewer in fullscreen mode
By default the file picker and the viewer are siblings of the fullscreen
element, so they are not visible when in fullscreen mode. It is not
possible to specify the parent nor to know when the file picker or the
viewer were actually opened, so for the time being they are reparented
if needed shortly after calling them.

This is just a temporary hack until it is possible to specify the parent
element, but that would require changes in the Vue library, the viewer
and the server dialogs, so this horrible code will have to do for now.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>