Use renegotiations to update publisher connections with the HPB
When the HPB supports the "update-sdp" feature sending an updated offer
for a publisher will propagate the renegotiation to the subscribers as
well. Therefore, in those cases there is no need to force a
reconnection, it is enough to send the new offer to trigger the
renegotiation (given that, in their latest versions, all the clients
will handle the updated offer using the current connection).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>