Fix unneeded checkbox name for checkbox not in group
If the "name" property is given to a "CheckboxRadioSwitch" the checkbox
is assumed to be part of a group, so "checked" needs to be an array
rather than a boolean. To prevent an error when mounting the component
and given that the checkbox is not used in a group the "name" property
was removed.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>