Fix loaded values in permissions dialog for participants
The "attendeePermissions" property provides the specific permissions for
the participant. However, due to the permission chain the actual
participant permissions may not be the default ones even if no custom
attendee permissions are set. The dialog should show the actual
participant permissions, so the "permissions" property should be used
instead.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>