Trigger event when same input track is set again with different state
As described in the TrackSinkMixin documentation when a different source
track is set the "_handleInputTrackEnabled()" method will not be called
even if the new track has different state than the previous track.
Therefore, in that case TrackToStream should trigger the "trackEnabled"
event, as otherwise the state change would not be noticed.

Currently this has no effect, as the input track was never set again
with a different enabled state, but it should make the code more robust.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>