Force enabled state before setting output track
The TrackEnabler node forces the enable state of the track when an input
track is set. However, this was done after setting the output track:
first the output track was set with the original state and then the
expected one was forced. Now the state is set before setting the output
track to ensure that the state will be the expected one from the start,
instead of having a transient state.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>