Set enabled state in model based on the current track state
Until now the enabled media state was based on the
"audioOn/audioOff/VideoOn/VideoOff" events triggered when enabling and
disabling the media. However, the enabled state was also explicitly
disabled when media was not available, for example after leaving a call.

Due to this, the enabled media state in the LocalMediaModel did not
necessarily reflect the current enabled media state in the tracks. For
example, if video was enabled in a previous call and a call is joined
now without camera, once a camera is selected again the track will be
enabled, but the video will be still disabled in the model until
something else causes it to change (like sending the current media state
to other participants).

To solve that the state in the LocalMediaModel is now set based on the
tracks, just like done in the peer connections, which should ensure
that the model reflects the local media state at all times. The enabled
state is still explicitly disabled when media is not available, but that
implicitly matches with the track state too (as if there is no track it
can not be enabled).

Moreover, media is no longer enabled and disabled by the model once
available; if media is available the model just reflects the enabled
state from the tracks, but it does not modify it. Therefore if media is
disabled when not available it may not be automatically disabled once
available, but this was not always working anyway since defe61742.

Finally, currently the local stream will contain at most one track of
each kind (audio or video), so the code assumes that and just checks the
first track of each kind, if any.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>