Fix participant list not refreshed when general permissions change
When the room or call permissions are set the permissions of all
participants could be modified. However, without HPB the participant
list was not updated in this case, so the participants may not be aware
of the change until up to 30 seconds, when the participant list was
automatically fetched again.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>