Fix media not disabled when not available
When media is not available the media was explicitly disabled to ensure
that it would not be unexpectedly enabled once available again. However,
due to the conditions inside "disableAudio()/disableVideo()" only the
local store was modified; the pipeline nodes that control whether tracks
are enabled or disabled were not modified, so when media was available
again in the same call it was automatically enabled back.

The conditions (and some supporting code) were needed in the past, but
since the introduction of the media pipeline it is possible to enable
and disable media even if there are no actual tracks (the tracks will
get enabled or disabled once they are started). Therefore the conditions
can be removed not only for disabling media but also for enabling it.

Besides that, disabling media when not available was done only when an
stream was updated; now it is also done when media is not available in
the initial stream.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>