Fix state not saved to local storage if call started without media
The enabled state of media is saved per conversation using the
conversation token. However, the token was not initialized when joining
a call without media, so either an empty token or a token from a
previous call was used. To solve this now the token is set not only if
media was started, but also if it failed to start.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>