Fix initial state not set when joining a call without media
The saved state was restored only when joining a call with media. This
was not a problem for audio or video, as in that case they were
automatically disabled due not having media available, but it caused the
virtual background state to be the default one or the one set in a
previous call.

Rather than also restoring the state after failing to start media the
state is now restored even before the media is tried to be initialized.
This ensures that the state will be the expected one once the media is
initialized, and with that prevent some unneeded changes back and forth
of the tracks caused by adjusting the audio, video and virtual
background once the tracks were already started.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>