Fix call flags update when track is disabled
When the call flags were updated after a renegotiation the senders were
checked to decide the flags to set. However, only the track attribute
was checked; when a track is disabled the track is nullified in the
sender and stored in "trackDisabled" instead, so the call flags were not
properly set if any of the tracks was disabled when a new one was added.

Note that when a local track is replaced the call flags are updated and,
in that case, the right flags are used, so sometimes this issue could be
masked by that.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>