Group on PHP level as MariaDB can not handle this correctly:
MySQL docs:
Reject queries for which the select list, HAVING condition, or ORDER BY
list refer to nonaggregated columns that are neither named in the GROUP BY
clause nor are functionally dependent on (uniquely determined by) GROUP BY columns.

MariaDB docs:
For SELECT ... GROUP BY queries, disallow SELECTing columns which are
not referred to in the GROUP BY clause, unless they are passed to an
aggregate function like COUNT() or MAX(). Produce a 1055 error.

Ref https://jira.mariadb.org/browse/MDEV-11588

Similarly Oracle seems to fail with the query before

Signed-off-by: Joas Schilling <coding@schilljs.com>