Generate actual constraints from the available tracks
The constraints are used in the caller code for things like setting the
current call flags. The constraints were modified by the
MediaDevicesSource and the MediaDevicesManager depending on the
available devices, but they did not reflect things like a node further
in the pipeline failing to set the track (for example, if the
VirtualBackground failed hard) and causing the final track to be null.
Basing the constraints on the tracks instead should address those
(corner) cases while behaving the same in the general cases.

Besides that, this will make possible to remove the constraints from
"MediaDevicesSource.start()" in a following commit.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>