Update call flags if needed after joining a call
When the local participant joins a call in most cases the call flags
already reflect the current media being published. However, if the
active media changed since the request to join the call was sent (for
example, if a track started during a forced reconnection, which could
happen if audio was disallowed and video allowed at the same time) the
call flags would need to be updated.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>