Fix joining call before leaving first in forced reconnections
When the HPB is not used forced reconnections just leave and then join
the call again. However, it was not waited for the call to be left first
before joining it again, so it could happen that the join request ended
being processed first and thus once the leave request was processed the
participant did not join again.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>