Add functions to configure the different parameters
Until now the script had to be edited to set the right configuration
parameters and then copied and pasted on the browser console; running
the script again required reloading the page and pasting it again, even
if the parameters where not modified. Now the constants were changed to
variables and helper functions were added to set them, so the script can
be pasted without changes and the configuration can be adjusted by
calling the helper functions. Moreover, as there is now an explicit
function to start the siege the script can be run several times and with
different parameters without having to reload the page and paste it
again.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>