Add CLI to run Talkbuchet
Talkbuchet is a JavaScript script meant to be loaded and run in a
browser. However, until now that required manually copying and pasting
the script in a browser console. This commit adds a Python script that
uses Selenium to launch a new browser, load Talkbuchet in it and control
it through the Python script.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>