Disconnect received videos by toggling the video directly in Janus
Originally the HPB required a renegotiation to block the received
videos. However, besides the standard WebRTC way, Janus also supports
directly toggling on and off the video without any renegotiation. This
is now exposed by the signaling server through the "selectStream"
message, so the video is now blocked using that approach instead.

The previous code to block the video using a renegotiation is still
kept, though, as it could be used without HPB once all the clients
support renegotiations.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>