Disable data channels for the SIP bridge subscriber
The SIP bridge publisher does not have data channels, so they need to be
explicitly disabled in the subscriber. Otherwise it would try to open
them, which would cause an endless loop of renegotiations, as after a
negotiation the data channels will still not be opened, which will
trigger a negotiation again.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>