Add explicit method to run asynchronous code
Executing "await someFunctionCall()" as synchronous code works as
expected in Chromium (the execution will not return until the awaited
function ended). However, it does not when explicit Promises are used
(it returns as soon as the main code finished, without waiting for the
Promise to be fulfilled).

WebDriver provides an explicit way to execute asynchronous functions and
signal when they finished to return, so this is used now for all
asynchronous code.

This will also make possible to add support for Firefox, as in Firefox
asynchronous functions executed as synchronous return immediately, so
they need to be explicitly executed as asynchronous.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>