Add helper class to get browser logs with BiDi protocol
WebDriver BiDi protocol provides bidirectional communication between a
WebDriver client and a web browser. In the original WebDriver protocol
the client sends requests, but the browser is not able to send requests
to the client on its own. Therefore, if a log was written by the browser
the browser could not notify the client, and the client will not notice
the log until it explicitly performs a new request. With the BiDi
protocol the client can subscribe to events in the browser, so the
browser will send them as soon as they happen.

A new helper class is introduced to get browser logs using the BiDi
protocol. Objects of this helper class will start their own thread to
ensure that the logs can be received as soon as they happen, as the main
Python thread could be be blocked waiting while a script is being
executed in the browser and that script can cause log messages to be
printed.

Unfortunately, Chromium does not implement yet getting the logs using
the BiDi protocol, so they still need to be got in the old way. Firefox,
on the other hand, only allows getting the logs through BiDi, so the
BiDi helper introduced here will be required to add support for Firefox.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>