Find default browser in Talkbuchet CLI
When Talkbuchet CLI is executed through the Talkbuchet runner the
browser available in the Docker image is automatically found, thus
removing the need to explicitly set it.

If Talkbuchet CLI is executed by other means, or if a remote Selenium
server is used, the browser still needs to be explicitly set.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>