Remove half-working code related to sieges with guest users
Automatically joining the room after getting the "hello" message is
needed only for guest users. However, sieges as guest users are not
possible, as joining the room a second time from the same browser causes
the first guest to be unregistered. Due to this the code to
automatically join the room is removed.

The method to join the room, however, is kept, as it will be used by the
"Virtual participant" mode to be introduced in a later commit.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>