Add virtual participant mode
In the virtual participant mode Talkbuchet can be used to add
participants to a call and thus perform load test in clients. Virtual
participants require much less resources than real participants in the
system running Talkbuchet, so it makes possible to add a lot more
participants than what would be possible with real participants.

Note, however, that a single browser can run just a single virtual
participant. Therefore, virtual participants are recommended to be used
through the CLI (the support will be introduced in a following commit).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>