Extract code specific to siege mode
In order to add support for the virtual participant mode in a following
commit the Talkbuchet wrapper class was split in a base class with the
common code and a subclass with the code specific to siege mode.
Similarly, the global helper functions were also split between the
common ones and the siege specific ones, which will only be defined when
"switchToSiegeMode()" is called (done by default).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>