Add real participant mode
The real participant mode is specific to Talkbuchet CLI and intended
only for development. This mode does not use Talkbuchet.js at all, and
just uses the real Talk app in the Nextcloud server to join the room
and, optionally, the call with a real participant.

Although it would provide the most accurate results in most cases it is
not suitable for load tests of the server or other clients, as the use
of real participants limits the number of participants that can be
created by the system running the CLI without saturating the CPU (unless
a remote Selenium server with plenty of available sessions is used).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>