Remove logs about simulcast stream not changed
The view assumes that the model will send a signaling message to set the
simulcast stream only when needed, so "selectSimulcastStream" is called
without verifying first whether it is needed or not. This can happen
quite often (for example, when a participant starts or stops speaking),
which can flood the logs in calls with a lot of participants, and
knowing that the stream was not changed is not very useful, so that
message is now removed.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>