Print peer ID along the peer object in logs
When the logs are checked live the peer objects can be expanded to know
which specific object they refer to. However, if the logs are exported
to a text file to be reviewed later there was no way to differentiate
which messages refer to which peer. Therefore, now the peer ID is
printed along the peer object.

The peer ID is also printed in the messages referring to the own peer
for convenience, although it is not strictly needed as in this case the
ID is implicit.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>