Fix missing Talk sidebar trigger in public share pages
Since the removal of SCSS files the "menu-people" icon can no longer be
used directly with a CSS class. However, it is available as a Vue
component; although it would be possible to directly render the Vue
component for the icon inside the button element the whole button is
moved to Vue instead, as the other approach requires more fighting with
the styles.

As the icon will be shown with a transparent background on the header,
which uses the primary color, the fill color of the icon needs to be
explicitly set to the primary text color.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>