Change the page title only after the route changed
This is an attempt to fix the problem that multiple links are
suggested by the browser when searching for a room by the name.
The assumption is that while the URL is being changed the page
title is recorded but we currently change the title to the new
name already.

Signed-off-by: Joas Schilling <coding@schilljs.com>