Split local streams and sent streams
Until now the stream locally shown was the stream sent to the other
peers. Now there are separate streams for each case to be able to adjust
them independently to work around specific browser behaviours.

Nevertheless, the sent stream is a single stream shared by all sender
peers, not a stream specific for each peer, so it is still generated in
the local media class, just like the local stream.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>