Fix voting for option 0
The problem is the entity had 0 as a default it would skip writing
it into the changeset of the query for the insert.
But the DB had a default of 1 so it would store a 1 afterwards in the db.

Having null as default on the Vote model fixes this, as then any number
is a change from the default, the mark-dirty happens and we add it to
the query.

Signed-off-by: Joas Schilling <coding@schilljs.com>