Fix remote screen shares not shown in sidebar
When the call view is shown in the sidebar there is no button to share
the screen. However, other participants can nevertheless open the call
from the main Talk UI and share their screen there, so remote screen
shares should be shown even if local ones can not be started.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>