Emit "leaveCall" event even if the request to leave the call fails
When a call is left the UI always changes to the "not-in-a-call" state.
However, if leaving the call fails the signaling did not emit the
"leaveCall" event, so the WebRTC connections were not stopped due to
leaving the call, nor they were left later when the signaling sent the
updated list of participants, as the participant did not leave the call
from the point of view of the server.

Even if leaving the call fails in the server the call should be locally
left when the user requests it, so now "leaveCall" is emitted even if
leaving the call failed.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>